+++ custom_components/soundtouchplus/const.py
@@ -22,6 +22,7 @@ DEFAULT_TIMEOUT = 15
 
 # custom service names.
 SERVICE_AUDIO_TONE_LEVELS = "audio_tone_levels"
+SERVICE_CLEAR_SOURCE_NOWPLAYINGSTATUS = "clear_source_nowplayingstatus"
 SERVICE_MUSICSERVICE_STATION_LIST = "musicservice_station_list"
 SERVICE_PLAY_CONTENTITEM = "play_contentitem"
 SERVICE_PLAY_HANDOFF = "play_handoff"
@@ -33,4 +34,5 @@ SERVICE_RECENT_LIST = "recent_list"
 SERVICE_REMOTE_KEYPRESS = "remote_keypress"
 SERVICE_SNAPSHOT_RESTORE = "snapshot_restore"
 SERVICE_SNAPSHOT_STORE = "snapshot_store"
+SERVICE_UPDATE_SOURCE_NOWPLAYINGSTATUS = "update_source_nowplayingstatus"
 SERVICE_ZONE_TOGGLE_MEMBER = "zone_toggle_member"